Refactor pin package into githubactions and container subpackages
Addresses structural and test strategy feedback:

- Extract GitHubActionsStrategy into internal/pin/githubactions/ with
  strategy, resolver, verifier, writer, and client
- Extract ContainerStrategy into internal/pin/container/ with strategy
  and rewrite
- Convert rewrite tests to golden file pattern using testdata/ input and
  .golden output fixtures for both ecosystems
- Export shared discovery helpers (ShouldSkipDir, IsSymlink, DedupeKey,
  IsWorkflowFile, IsCommitSHA) from the pin package

Quality and security improvements found during review:

- Fix silent data loss: check f.Close() errors in both file writers
- Fix injection: validate Name/VersionTag for newlines in container writer
- Fix cache race: use singleflight to deduplicate concurrent ls-remote calls
- Fix context cancellation: select on ctx.Done() when acquiring semaphore
- Validate exclude patterns up front instead of silently ignoring bad globs
- Skip action refs with no @version in composite action scanning
- Hoist fromRe regex to package-level var (was recompiled per Dockerfile)
- Add debug logging for unparseable files in container discovery
- Modernize sort.Slice to slices.SortFunc per project conventions
