+++ internal/pin/container/testdata/dockerfile_basic.golden
@@ -1,0 +1,2 @@
+FROM alpine:3.19@sha256:a8560b36e8b8210634f77d9f7f9efd7ffa463e380b75e2e74aff4511df3ef88c
+RUN echo hello
