+++ internal/pin/container/testdata/dockerfile_basic.input
@@ -1,0 +1,2 @@
+FROM alpine:3.19
+RUN echo hello
