+++ internal/pin/container/testdata/dockerfile_multistage.golden
@@ -1,0 +1,5 @@
+FROM golang:1.23@sha256:1111111111111111111111111111111111111111111111111111111111111111 AS builder
+COPY . .
+RUN go build -o /app .
+FROM alpine:3.19@sha256:2222222222222222222222222222222222222222222222222222222222222222
+COPY --from=builder /app /app
