+++ internal/pin/container/testdata/dockerfile_platform.input
@@ -1,0 +1,2 @@
+FROM --platform=linux/amd64 node:22-slim
+RUN echo hello
