+++ internal/pin/container/testdata/dockerfile_replace_digest.golden
@@ -1,0 +1,1 @@
+FROM alpine:3.19@sha256:1111111111111111111111111111111111111111111111111111111111111111
