+++ internal/pin/container/testdata/dockerfile_replace_digest.input
@@ -1,0 +1,1 @@
+FROM alpine:3.19@sha256:0000000000000000000000000000000000000000000000000000000000000000
