+++ internal/pin/container/testdata/workflow_containers.golden.yml
@@ -1,0 +1,10 @@
+name: CI
+on: push
+jobs:
+  test:
+    container: postgres:16@sha256:1111111111111111111111111111111111111111111111111111111111111111
+    services:
+      redis:
+        image: redis:7@sha256:2222222222222222222222222222222222222222222222222222222222222222
+    steps:
+      - uses: docker://alpine:3.19@sha256:3333333333333333333333333333333333333333333333333333333333333333
