+++ internal/pin/githubactions/testdata/already_pinned.golden.yml
