+++ internal/pin/githubactions/testdata/mixed.yml
