+++ internal/pin/githubactions/testdata/quoted.golden.yml
@@ -1,0 +1,8 @@
+name: CI
+on: push
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: "actions/checkout@abc123def456abc123def456abc123def456abc1" # v4.2.2
+      - uses: 'actions/setup-go@def456abc123def456abc123def456abc123def4' # v5.4.0
