+++ internal/pin/githubactions/testdata/quoted.yml
